# api/openapi.yaml
openapi: 3.0.3
info:
  title: Weave Ontology API
  version: 1.0.0
  description: |
    Weave is a single-machine Ontology Layer engine inspired by Palantir Foundry OSv2.
    This document describes its REST API: ontology metadata, object retrieval and
    search, ObjectSet composition, aggregation, link traversal, action execution,
    and authentication.
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development (port configurable via WEAVE_PORT).
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        RS256-signed JWT issued via /api/auth/login. Send as
        `Authorization: Bearer <token>`.
    apiKey:
      type: apiKey
      in: header
      name: X-Weave-Api-Key
      description: Reserved for future API-key-based service-to-service auth.
  responses:
    BadRequestError:
      description: The request was malformed or failed validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnauthorizedError:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFoundError:
      description: The requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  parameters:
    OntologyApiName:
      in: path
      name: ontologyApiName
      required: true
      description: API name (or RID) of the ontology.
      schema:
        type: string
...
    ObjectType:
      in: path
      name: objectType
      required: true
      description: API name of the object type.
      schema:
        type: string
    PrimaryKey:
      in: path
      name: primaryKey
      required: true
      description: Primary key of the object.
      schema:
        type: string
...
  schemas:
    ErrorResponse:
      type: object
      description: Palantir-style API error envelope.
      required: [errorCode, errorName, errorInstanceId, parameters]
      properties:
        errorCode:
          type: string
          example: NOT_FOUND
        errorName:
          type: string
          example: OntologyNotFound
        errorInstanceId:
          type: string
          format: uuid
        parameters:
          type: object
          additionalProperties:
            type: string
...
    WireObject:
      type: object
      description: |
        Palantir V2 wire format for an object: properties are flattened at the
        top level alongside __rid, __primaryKey, and __apiName.
      additionalProperties: true
      properties:
        __rid:
          type: string
        __primaryKey: {}
        __apiName:
          type: string
    ObjectPage:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/WireObject"
        nextPageToken:
          type: string
        totalCount:
          type: string
    ObjectHistoryEntry:
      type: object
      description: |
        A single revision of an object's state. Tier 2.3 records one entry
        for each CREATE / MODIFY / DELETE applied to a primary key.
      properties:
        id:
          type: string
          description: Server-generated history row id (UUID).
        objectTypeRid:
          type: string
        primaryKey:
          type: string
        version:
          type: integer
          format: int64
          description: Monotonically increasing per-primary-key version.
        prevState:
          description: |
            JSONB snapshot of the object's state BEFORE the edit. Null for
            CREATE rows.
          type: object
          additionalProperties: true
          nullable: true
        newState:
          description: |
            JSONB snapshot of the object's state AFTER the edit. Null for
            DELETE rows.
          type: object
          additionalProperties: true
          nullable: true
        editType:
          type: string
          enum: [CREATE, MODIFY, DELETE]
        actionLogRid:
          type: string
          description: Optional RID of the action log row that produced this edit.
        userId:
          type: string
        recordedAt:
          type: string
          format: date-time
    ObjectHistoryResponse:
      type: object
      properties:
        history:
          type: array
          items:
            $ref: "#/components/schemas/ObjectHistoryEntry"
        totalVersions:
          type: integer
          format: int64
          description: Total number of history rows recorded for this object.
    SearchObjectsRequest:
      type: object
      properties:
        where:
          type: object
          additionalProperties: true
          description: WhereClause filter (see pkg/oss/where).
        select:
          type: array
          items:
            type: string
        pageSize:
          type: integer
        pageToken:
          type: string
        orderBy:
          type: object
          properties:
            fields:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  direction:
                    type: string
                    enum: [asc, desc]
...
security:
  - bearerAuth: []
paths:
...
  /api/v2/ontologies/{ontologyApiName}/objects/{objectType}:
    get:
      operationId: listObjects
      summary: List objects of a given type
      description: Paginated, optionally ordered list of objects.
      tags: [Objects]
      parameters:
        - $ref: "#/components/parameters/OntologyApiName"
        - $ref: "#/components/parameters/ObjectType"
        - in: query
          name: pageSize
          schema:
            type: integer
        - in: query
          name: pageToken
          schema:
            type: string
        - in: query
          name: orderBy
          schema:
            type: string
      responses:
        "200":
          description: Object page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObjectPage"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /api/v2/ontologies/{ontologyApiName}/objects/{objectType}/{primaryKey}:
    get:
      operationId: getObject
      summary: Get a single object by primary key
      tags: [Objects]
      parameters:
        - $ref: "#/components/parameters/OntologyApiName"
        - $ref: "#/components/parameters/ObjectType"
        - $ref: "#/components/parameters/PrimaryKey"
      responses:
        "200":
          description: The object.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WireObject"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /api/v2/ontologies/{ontologyApiName}/objects/{objectType}/search:
    post:
      operationId: searchObjects
      summary: Search objects with a where clause
      tags: [Objects]
      parameters:
        - $ref: "#/components/parameters/OntologyApiName"
        - $ref: "#/components/parameters/ObjectType"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchObjectsRequest"
      responses:
        "200":
          description: Object page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObjectPage"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /api/v2/ontologies/{ontologyApiName}/objects/{objectType}/{primaryKey}/links/{linkType}:
    get:
      operationId: listLinkedObjects
      summary: List objects linked to a source object via a link type
      tags: [Objects]
      parameters:
        - $ref: "#/components/parameters/OntologyApiName"
        - $ref: "#/components/parameters/ObjectType"
        - $ref: "#/components/parameters/PrimaryKey"
        - in: path
          name: linkType
          required: true
          schema:
            type: string
        - in: query
          name: direction
          schema:
            type: string
            enum: [forward, reverse]
        - in: query
          name: pageSize
          schema:
            type: integer
        - in: query
          name: pageToken
          schema:
            type: string
      responses:
        "200":
          description: Linked object page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObjectPage"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /api/v2/ontologies/{ontologyApiName}/objects/{objectType}/{primaryKey}/history:
    get:
      operationId: getObjectHistory
      summary: Get the change history for a single object
      description: |
        Returns the most recent revisions of an object's state, ordered by
        version DESC, plus the total number of versions ever recorded. Each
        history row carries a pre/post-state snapshot, the edit type
        (CREATE / MODIFY / DELETE), the user that triggered the change, and
        the recorded timestamp.
      tags: [Objects]
      parameters:
        - $ref: "#/components/parameters/OntologyApiName"
        - $ref: "#/components/parameters/ObjectType"
        - $ref: "#/components/parameters/PrimaryKey"
        - in: query
          name: limit
          description: Maximum number of history rows to return (default 50, max 500).
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 50
      responses:
        "200":
          description: Object history page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObjectHistoryResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /api/v2/ontologies/{ontologyApiName}/links/{linkTypeApiName}:
    post:
      operationId: createLink
      summary: Upsert a single many-to-many link edge
      description: |
        Creates or updates a single link edge between two objects via a
        MANY_TO_MANY link type. Idempotent — calling with identical
        arguments is a no-op. Returns 404 if the link type does not exist
        or is declared with a non-M2M cardinality (M2O/O2M links must be
        mutated via object updates, not the link CRUD API).
      tags: [Objects]
      parameters:
        - $ref: "#/components/parameters/OntologyApiName"
        - in: path
          name: linkTypeApiName
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sourcePk, targetPk]
              properties:
                sourcePk:
                  type: string
                  description: Primary key of the source object
                targetPk:
                  type: string
                  description: Primary key of the target object
                properties:
                  type: object
                  additionalProperties: true
                  description: Edge properties, stored as JSONB
      responses:
        "201":
          description: Link edge created or updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  linkType:
                    type: string
                  sourcePrimaryKey:
                    type: string
                  targetPrimaryKey:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
    delete:
      operationId: deleteLink
      summary: Delete a single many-to-many link edge
      description: |
        Removes a single edge from a MANY_TO_MANY link type. Idempotent —
        deleting a non-existent edge returns 200 with status=deleted.
      tags: [Objects]
      parameters:
        - $ref: "#/components/parameters/OntologyApiName"
        - in: path
          name: linkTypeApiName
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sourcePk, targetPk]
              properties:
                sourcePk:
                  type: string
                targetPk:
                  type: string
      responses:
        "200":
          description: Link edge deleted (or did not exist).
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  linkType:
                    type: string
                  sourcePrimaryKey:
                    type: string
                  targetPrimaryKey:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
...